requirements.txt

Werkzeug==2.3.3
python-dotenv==0.19.1
google-generativeai==0.3.1
googletrans==3.1.0a0
SpeechRecognition==3.8.1
gTTS==2.2.3
pydub==0.25.1
PyAudio==0.2.11
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
itsdangerous==2.0.1
requests==2.26.0
twilio==7.9.0
Flask==2.1.0
python-dotenv==0.19.2
requests==2.26.0
sendgrid==6.9.7
Pillow==10.4.0